Implementado o grande método para iniciar partida!

+++ back-end/questoes-java/src/main/java/com/lipe/questoes_br/Repository/PerguntaRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.lipe.questoes_br.Entity.Pergunta;
 
-public interface PerguntaRepository extends JpaRepository<Pergunta, Integer> {
-    
-    @Query(value="SELECT * FROM pergunta order by Random() LIMIT ?1", nativeQuery= true)
+public interface PerguntaRepository extends JpaRepository<Pergunta, Long> {
+
+    @Query(value = "SELECT * FROM pergunta order by Random() LIMIT ?1", nativeQuery = true)
     List<Pergunta> coletarPerguntas(int quantidade);
-}   
+}
